+++ src/ll/xist/ns/doc.py
@@ -79,7 +79,7 @@ def getdoc(thing, format):
 			realthing = thing
 			while hasattr(realthing, "__wrapped__"):
 				realthing = realthing.__wrapped__
-			for ref in node.walknode(pyref):
+			for ref in node.walknodes(pyref):
 				if u"module" not in ref.attrs:
 					ref[u"module"] = _getmodulename(realthing)
 					if u"class_" not in ref.attrs:
@@ -90,17 +90,17 @@ def getdoc(thing, format):
 			# Use the original method instead of the decorator
 			while hasattr(thing, "__wrapped__"):
 				thing = thing.__wrapped__
-			for ref in node.walknode(pyref):
+			for ref in node.walknodes(pyref):
 				if u"module" not in ref.attrs:
 					ref[u"module"] = _getmodulename(thing)
 		elif inspect.isclass(thing):
-			for ref in node.walknode(pyref):
+			for ref in node.walknodes(pyref):
 				if u"module" not in ref.attrs:
 					ref[u"module"] = _getmodulename(thing)
 					if u"class_" not in ref.attrs:
 						ref[u"class_"] = thing.__name__
 		elif inspect.ismodule(thing):
-			for ref in node.walknode(pyref):
+			for ref in node.walknodes(pyref):
 				if u"module" not in ref.attrs:
 					ref[u"module"] = thing.__name__
 		return node