+++ src/ll/xist/scripts/tld2xsc.py
@@ -35,7 +35,7 @@ def tld2xnd(stream, shareattrs=None):
 	node = parsers.parsestream(stream, prefixes={None: tld})
 
 	# get and convert the taglib object
-	xnd = node.walknode(tld.taglib)[0].asxnd()
+	xnd = node.walknodes(tld.taglib)[0].asxnd()
 
 	if shareattrs=="dupes":
 		xnd.shareattrs(False)